Revive backend-integration
