+++ tests/integration/relations/pgbouncer_provider/application-charm/metadata.yaml
@@ -8,7 +8,7 @@ summary: |
   only for testing of the libs in this repository.
 
 series:
-  - focal
+  - jammy
 
 requires:
   first-database:
