+++ tests/integration/relations/pgbouncer_provider/application-charm/requirements.txt
@@ -1,1 +1,1 @@
-ops==1.5.4
+ops==2.1.1
